Fix issue with lazy load only loading 1 image
Looks like owl carousel can't load more than 1 image per slide. I need a slide that is a grid of many images, and for all those images on the slide to be loaded when that slide becomes active. The previous behavior was it would load only the 1st image encountered on a slide. This change causes it to load all the images on the active slide.